Added ability for EventDispacher to connect to member functions
What a nightmare this was... If only member functions could be cast to void*

+++ src/core/graphics/GPUMemory.cpp
@@ -10,7 +10,7 @@ GPUMemory::~GPUMemory() {
 }
 
 GPUMemory* GPUMemory::create(const GPUMemoryConfiguration& gpuMemoryConfiguration) {
-	float printAllocSize = gpuMemoryConfiguration.size;
+	float printAllocSize = (float)gpuMemoryConfiguration.size;
 	std::string printSizeUnit = "Bytes";
 
 	if (printAllocSize >= 1024) {
